+++ src/components/board/BoardListComponent.js
@@ -8,6 +8,25 @@ class BoardListComponent {
     get inputListName() { return $('[data-testid="list-name-textarea"]'); }
     get addListToBoardButton() { return $('[data-testid="list-composer-add-list-button"]'); }
     listElement(listTitle) { return $(`h2[data-testid="list-name"]=${listTitle}`); }
+    get inputCardName() { return $('[data-testid="list-card-composer-textarea"]'); }
+    get addCardToListButton() { return $('[data-testid="list-card-composer-add-card-button"]'); }
+    cardElement(cardTitle) { return $(`a[data-testid="card-name"]=${cardTitle}`); }
+
+    async enterCardName(cardName) {
+        await this.inputCardName.waitForDisplayed();
+        await this.inputCardName.click();
+        await this.inputCardName.setValue(cardName);
+    }
+
+    async clickAddCardToList() {
+        await this.addCardToListButton.waitForDisplayed();
+        await this.addCardToListButton.click();
+    }
+
+    async isCardDisplayed(cardTitle) {
+        const card = this.cardElement(cardTitle);
+        return await card.isExisting() && card.isDisplayed();
+    }
 
     async enterListName(listName) {
         await this.inputListName.waitForDisplayed();
