--- src/pages/CardMangementPage.js
@@ -1,8 +1,0 @@
-const cardManagementComponent = require('../components/CardManagementComponent');
-
-class CardManagmentPage {
-    constructor() {
-        this.cardManagement = cardManagementComponent;
-    }
-}
-module.exports = new CardManagmentPage();
